+++ examples/Connect-AvptElements.ps1
@@ -29,3 +29,5 @@ Connect-AvptElements
 Get-AvptCustomer -All
 Get-AvptBackupOverview -CustomerId '00000000-0000-0000-0000-000000000000'
 Get-AvptScanProfile -CustomerId '00000000-0000-0000-0000-000000000000'
+Get-AvptBaseline -All
+Get-AvptRiskRule -CustomerId '00000000-0000-0000-0000-000000000000' -TenantId '11111111-1111-1111-1111-111111111111'