+++ src/AvePoint.Elements/Public/Get-AvptCustomer.ps1
@@ -22,10 +22,13 @@ function Get-AvptCustomer {
 
     $maps = Get-AvptEnumMap
     foreach ($item in $items) {
-        ConvertTo-AvptFriendlyObject -InputObject $item -TypeName 'AvePoint.Elements.Customer' -EnumMap @{
+        $customer = ConvertTo-AvptFriendlyObject -InputObject $item -TypeName 'AvePoint.Elements.Customer' -EnumMap @{
             JobStatus      = $maps.JobStatus
             ManagementMode = $maps.ManagementMode
         }
+        $customer | Add-Member -NotePropertyName TenantCount -NotePropertyValue @($item.tenants).Count
+        $customer | Add-Member -NotePropertyName TenantNames -NotePropertyValue (@($item.tenants | ForEach-Object { $_.name }) -join ', ')
+        $customer.tenants = @($item.tenants | ForEach-Object { ConvertTo-AvptNestedObject -InputObject $_ -TypeName 'AvePoint.Elements.CustomerTenant' })
+        $customer
     }
 }
-